automation/requirements.txt
appdirs (1.4.3)
packaging (16.8)
pexpect (4.2.1)
pip (9.0.1)
pkg-resources (0.0.0)
ptyprocess (0.5.1)
py (1.4.33)
pyio (0.1.1)
pyparsing (2.2.0)
pytest (3.0.7)
PyYAML (3.12)
setuptools (35.0.2)
six (1.10.0)
wheel (0.30.0a0)
